+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Add Samsung browser.
 - Add The Knowledge AI bot to the bot list.
 - Add Go HttpClient to the bot list.
-- Fix Microsoft Edge detection on iOS & Android
+- Fix Microsoft Edge detection on Android and iOS.
+- Fix MicroMessenger detection on Android
 
 ## 4.0.0